+++ Nt/386/include/lib9.h
@@ -53,7 +53,7 @@ typedef unsigned int	uint;
 typedef unsigned long	ulong;
 typedef signed char	schar;
 typedef unsigned short	ushort;
-typedef unsigned short	Rune;
+typedef unsigned int Rune;
 typedef __int64		vlong;
 typedef unsigned __int64		uvlong;
 typedef unsigned int u32int;
